Yse double for value calculations and update related methods

+++ cpp09/ex00/BitcoinExchange.hpp
@@ -6,7 +6,7 @@
 /*   By: edegraev <edegraev@student.forty2.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/03 14:19:18 by edegraev          #+#    #+#             */
-/*   Updated: 2024/12/06 09:54:27 by edegraev         ###   ########.fr       */
+/*   Updated: 2024/12/06 10:10:34 by edegraev         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,14 +21,14 @@
 class BitcoinExchange
 {
 private:
-    std::map<std::string, float> _btcHistory; // double ?
+    std::map<std::string, double> _btcHistory;
     
     void parseData(std::string const &line, bool isUser);
-    void checkValue(std::string data, float nbBtc);
+    void checkValue(std::string data, double nbBtc);
     bool isDate(std::string date);
     bool isDatePrevious(std::string dateBtc, std::string dateUser);
     
-    static float stof(std::string const &str);
+    static double stod(std::string const &str);
     static int stoi(std::string const &str);
 
 public:
